Extract utility functions out of gradients_test.cc so they be used in other tests.
PiperOrigin-RevId: 342152412
